split tests into own files
