+++ tests/Testcontainers.Dapr.Tests/SmokeTests.cs
@@ -1,0 +1,38 @@
+namespace Testcontainers.Dapr;
+
+public sealed class SmokeTest : IAsyncLifetime
+{
+    private DaprContainer _daprContainer;
+
+    public Task InitializeAsync()
+    {
+        _daprContainer = new DaprBuilder()
+            .WithAppId("test-app")
+            .WithLogLevel("debug")
+            .Build();
+
+        return _daprContainer.StartAsync();
+    }
+
+    public Task DisposeAsync()
+    {
+        return _daprContainer.DisposeAsync().AsTask();
+    }
+
+    [Fact]
+    [Trait(nameof(DockerCli.DockerPlatform), nameof(DockerCli.DockerPlatform.Linux))]
+    public async Task CheckDaprSideCarIsHealthy()
+    {            
+         // Given
+        using var client = new DaprClientBuilder()
+            .UseHttpEndpoint(_daprContainer.GetHttpAddress())
+            .UseGrpcEndpoint(_daprContainer.GetGrpcAddress())
+            .Build();
+
+        // When
+        var healthy = await client.CheckHealthAsync();
+
+        // Then
+        Assert.True(healthy);
+    }
+}
